Reorganize panel toolbar buttons
